Seperate methods to import them as a module
